requirements.txt
basicsr
facexlib
gfpgan
torch
torchvision
numpy
opencv-python
pillow
ttkbootstrap
